+++ buildkite/resource_pipeline.go
@@ -61,11 +61,6 @@ type TeamPipelineNode struct {
 	Team        TeamNode
 }
 
-type PipelineTeamAssignmentInput struct {
-	Id          graphql.ID     `json:"id"`
-	AccessLevel graphql.String `json:"accessLevel"`
-}
-
 // resourcePipeline represents the terraform pipeline resource schema
 func resourcePipeline() *schema.Resource {
 	return &schema.Resource{
@@ -331,11 +326,6 @@ func CreatePipeline(ctx context.Context, d *schema.ResourceData, m interface{}) 
 	var err error
 
 	teamPipelines := getTeamPipelinesFromSchema(d)
-	var mutation struct {
-		PipelineCreate struct {
-			Pipeline PipelineNode
-		} `graphql:"pipelineCreate(input: {allowRebuilds: $allow_rebuilds, branchConfiguration: $branch_configuration, cancelIntermediateBuilds: $cancel_intermediate_builds, cancelIntermediateBuildsBranchFilter: $cancel_intermediate_builds_branch_filter, defaultBranch: $default_branch, defaultTimeoutInMinutes: $default_timeout_in_minutes, maximumTimeoutInMinutes: $maximum_timeout_in_minutes, description: $desc, name: $name, organizationId: $org, repository: {url: $repository_url}, skipIntermediateBuilds: $skip_intermediate_builds, skipIntermediateBuildsBranchFilter: $skip_intermediate_builds_branch_filter, steps: {yaml: $steps}, teams: $teams, tags: $tags})"`
-	}
 
 	teamsData := make([]PipelineTeamAssignmentInput, 0)
 	for _, team := range teamPipelines {
@@ -346,53 +336,47 @@ func CreatePipeline(ctx context.Context, d *schema.ResourceData, m interface{}) 
 		}
 		teamsData = append(teamsData, PipelineTeamAssignmentInput{
 			Id:          teamID,
-			AccessLevel: graphql.String(team.AccessLevel),
+			AccessLevel: team.AccessLevel,
 		})
 	}
 
-	vars := map[string]interface{}{
-		"allow_rebuilds":                           graphql.Boolean(d.Get("allow_rebuilds").(bool)),
-		"branch_configuration":                     graphql.String(d.Get("branch_configuration").(string)),
-		"cancel_intermediate_builds":               graphql.Boolean(d.Get("cancel_intermediate_builds").(bool)),
-		"cancel_intermediate_builds_branch_filter": graphql.String(d.Get("cancel_intermediate_builds_branch_filter").(string)),
-		"default_branch":                           graphql.String(d.Get("default_branch").(string)),
-		"default_timeout_in_minutes":               graphql.Int(d.Get("default_timeout_in_minutes").(int)),
-		"maximum_timeout_in_minutes":               graphql.Int(d.Get("maximum_timeout_in_minutes").(int)),
-		"desc":                                     graphql.String(d.Get("description").(string)),
-		"name":                                     graphql.String(d.Get("name").(string)),
-		"org":                                      client.organizationId,
-		"repository_url":                           graphql.String(d.Get("repository").(string)),
-		"skip_intermediate_builds":                 graphql.Boolean(d.Get("skip_intermediate_builds").(bool)),
-		"skip_intermediate_builds_branch_filter":   graphql.String(d.Get("skip_intermediate_builds_branch_filter").(string)),
-		"steps":                                    graphql.String(d.Get("steps").(string)),
-		"teams":                                    teamsData,
-		"tags":                                     getTagsFromSchema(d),
+	input := PipelineCreateInput{
+		AllowRebuilds:                        d.Get("allow_rebuilds").(bool),
+		BranchConfiguration:                  d.Get("branch_configuration").(string),
+		CancelIntermediateBuilds:             d.Get("cancel_intermediate_builds").(bool),
+		CancelIntermediateBuildsBranchFilter: d.Get("cancel_intermediate_builds_branch_filter").(string),
+		DefaultBranch:                        d.Get("default_branch").(string),
+		DefaultTimeoutInMinutes:              d.Get("default_timeout_in_minutes").(int),
+		MaximumTimeoutInMinutes:              d.Get("maximum_timeout_in_minutes").(int),
+		Description:                          d.Get("description").(string),
+		Name:                                 d.Get("name").(string),
+		OrganizationId:                       client.organizationId,
+		Repository:                           PipelineRepositoryInput{Url: d.Get("repository").(string)},
+		SkipIntermediateBuilds:               d.Get("skip_intermediate_builds").(bool),
+		SkipIntermediateBuildsBranchFilter:   d.Get("skip_intermediate_builds_branch_filter").(string),
+		Steps:                                PipelineStepsInput{Yaml: d.Get("steps").(string)},
+		Teams:                                teamsData,
+		Tags:                                 getTagsFromSchema(d),
 	}
 
-	log.Printf("Creating pipeline %s ...", vars["name"])
+	log.Printf("Creating pipeline %s ...", d.Get("name"))
 
 	// If the cluster_id key is present in the mutation, GraphQL expects a valid ID.
 	// Check if cluster_id exists in the configuration before adding to mutation.
 	if clusterID, ok := d.GetOk("cluster_id"); ok {
-		var mutationWithClusterID struct {
-			PipelineCreate struct {
-				Pipeline PipelineNode
-			} `graphql:"pipelineCreate(input: {allowRebuilds: $allow_rebuilds, branchConfiguration: $branch_configuration, cancelIntermediateBuilds: $cancel_intermediate_builds, cancelIntermediateBuildsBranchFilter: $cancel_intermediate_builds_branch_filter, clusterId: $cluster_id, defaultBranch: $default_branch, defaultTimeoutInMinutes: $default_timeout_in_minutes, description: $desc, maximumTimeoutInMinutes: $maximum_timeout_in_minutes, name: $name, organizationId: $org, repository: {url: $repository_url}, skipIntermediateBuilds: $skip_intermediate_builds, skipIntermediateBuildsBranchFilter: $skip_intermediate_builds_branch_filter, steps: {yaml: $steps}, teams: $teams, tags: $tags})"`
-		}
-		vars["cluster_id"] = graphql.ID(clusterID.(string))
-		err = client.graphql.Mutate(context.Background(), &mutationWithClusterID, vars)
-		mutation.PipelineCreate.Pipeline = mutationWithClusterID.PipelineCreate.Pipeline
-	} else {
-		err = client.graphql.Mutate(context.Background(), &mutation, vars)
+		id := clusterID.(string)
+		input.ClusterId = &id
 	}
+
+	response, err := createPipeline(client.genqlient, input)
 
 	if err != nil {
 		log.Printf("Unable to create pipeline %s", d.Get("name"))
 		return diag.FromErr(err)
 	}
-	log.Printf("Successfully created pipeline with id '%s'.", mutation.PipelineCreate.Pipeline.ID)
+	log.Printf("Successfully created pipeline with id '%s'.", response.PipelineCreate.Pipeline.Id)
 
-	updatePipelineResource(d, &mutation.PipelineCreate.Pipeline)
+	updatePipelineResourceFromCreate(d, response.PipelineCreate.Pipeline)
 
 	pipelineExtraInfo, err := updatePipelineExtraInfo(d, client)
 	if err != nil {
@@ -804,6 +788,36 @@ func deleteTeamPipelines(teamPipelines []TeamPipelineNode, client *Client) error
 	}
 
 	return nil
+}
+
+func updatePipelineResourceFromCreate(d *schema.ResourceData, pipeline createPipelinePipelineCreatePipelineCreatePayloadPipeline) {
+	d.SetId(string(pipeline.Id))
+	d.Set("allow_rebuilds", bool(pipeline.AllowRebuilds))
+	d.Set("branch_configuration", string(pipeline.BranchConfiguration))
+	d.Set("default_timeout_in_minutes", int(pipeline.DefaultTimeoutInMinutes))
+	d.Set("maximum_timeout_in_minutes", int(pipeline.MaximumTimeoutInMinutes))
+	d.Set("cancel_intermediate_builds", bool(pipeline.CancelIntermediateBuilds))
+	d.Set("cancel_intermediate_builds_branch_filter", string(pipeline.CancelIntermediateBuildsBranchFilter))
+	d.Set("cluster_id", string(pipeline.Cluster.Id))
+	d.Set("default_branch", string(pipeline.DefaultBranch))
+	d.Set("description", string(pipeline.Description))
+	d.Set("name", string(pipeline.Name))
+	d.Set("repository", string(pipeline.Repository.Url))
+	d.Set("skip_intermediate_builds", bool(pipeline.SkipIntermediateBuilds))
+	d.Set("skip_intermediate_builds_branch_filter", string(pipeline.SkipIntermediateBuildsBranchFilter))
+	d.Set("slug", string(pipeline.Slug))
+	d.Set("steps", string(pipeline.Steps.Yaml))
+	d.Set("webhook_url", string(pipeline.WebhookURL))
+
+	teams := make([]map[string]interface{}, len(pipeline.Teams.Edges))
+	for i, id := range pipeline.Teams.Edges {
+		team := map[string]interface{}{
+			"slug":         string(id.Node.Team.Slug),
+			"access_level": string(id.Node.AccessLevel),
+		}
+		teams[i] = team
+	}
+	d.Set("team", teams)
 }
 
 // updatePipelineResource updates the terraform resource data for the pipeline resource